requirements.txt
amqp==2.5.1
asn1crypto==1.1.0
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
billiard==3.6.1.0
boto==2.49.0
boto3==1.9.251
botocore==1.12.251
cached-property==1.5.1
celery==4.3.0
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
CodeConvert==3.0.2
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
cryptography==2.7
cssutils==1.0.2
defusedxml==0.6.0
Django==2.2.6
django-admin==2.0.0
django-allauth==0.40.0
django-celery-beat==1.5.0
django-cors-headers==3.1.1
django-excel-response2==2.0.8
django-inlinecss==0.3.0
django-models-ext==1.1.9
django-redis==4.10.0
django-rest-swagger==2.1.2
django-short-url==1.1.4
django-six==1.0.4
django-storages==1.7.2
django-timezone-field==3.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
Flask==1.1.1
furl==2.1.0
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
idna==2.7
importlib-metadata==0.23
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==2.6.0
kombu==4.6.5
locust==0.0
locustio==0.11.0
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.2
mysqlclient==1.4.4
oauthlib==3.1.0
openapi-codec==1.3.2
orderedmultidict==1.0.1
packaging==19.2
paramiko==2.6.0
Pillow==6.2.0
pluggy==0.13.0
py==1.8.0
pycparser==2.19
pyee==6.0.0
PyJWT==1.7.1
PyNaCl==1.3.0
pynliner==0.8.0
pyparsing==2.4.2
pytest==5.2.1
pytest-cov==2.8.1
pytest-django==3.5.1
python-crontab==2.4.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.3
PyYAML==3.13
pyzmq==18.1.0
redis==3.3.11
requests==2.20.1
requests-oauthlib==1.2.0
s3transfer==0.2.1
screen==1.0.1
shortuuid==0.5.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
texttable==0.9.1
TimeConvert==1.5.0
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.24.3
vine==1.3.0
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.16.0
xlwt==1.3.0
zipp==0.6.0